+++ src/lib.rs
@@ -1,2 +1,5 @@
-pub mod time_map;
-pub mod director;
+#[macro_use]
+extern crate lazy_static;
+
+pub mod time;
+pub mod midi;
